+++ astropy/table/table.py
@@ -2086,6 +2086,91 @@ class Table(object):
 
         return groups.table_group_by(self, keys)
 
+    def to_pandas(self):
+        """
+        Return a :class:`pandas.DataFrame` instance
+
+        Returns
+        -------
+        dataframe : :class:`pandas.DataFrame`
+            A pandas :class:`pandas.DataFrame` instance
+
+        Raises
+        ------
+        ImportError
+            If pandas is not installed
+        ValueError
+            If the Table contains mixin or multi-dimensional colunns
+        """
+        from pandas import DataFrame
+
+        if self.has_mixin_columns:
+            raise ValueError("Cannot convert a table with mixin columns to a pandas DataFrame")
+
+        if any(getattr(col, 'ndim', 1) > 1 for col in self.columns.values()):
+            raise ValueError("Cannot convert a table with multi-dimensional columns to a pandas DataFrame")
+
+        out = OrderedDict()
+
+        for name, column in self.columns.items():
+            if isinstance(column, MaskedColumn):
+                if column.dtype.kind in ['i', 'u']:
+                    out[name] = column.astype(float).filled(np.nan)
+                elif column.dtype.kind in ['f', 'c']:
+                    out[name] = column.filled(np.nan)
+                else:
+                    out[name] = column.astype(np.object).filled(np.nan)
+            else:
+                out[name] = column
+
+        return DataFrame(out)
+
+    @classmethod
+    def from_pandas(cls, dataframe):
+        """
+        Create a `Table` from a :class:`pandas.DataFrame` instance
+
+        Parameters
+        ----------
+        dataframe : :class:`pandas.DataFrame`
+            The pandas :class:`pandas.DataFrame` instance
+
+        Returns
+        -------
+        table : `Table`
+            A `Table` (or subclass) instance
+        """
+
+        out = OrderedDict()
+
+        for name in dataframe.columns:
+            column = dataframe[name]
+            mask = np.array(column.isnull())
+            data = np.array(column)
+
+            if data.dtype.kind == 'O':
+                # If all elements of an object array are string-like or np.nan
+                # then coerce back to a native numpy str/unicode array.
+                string_types = six.string_types
+                if six.PY3:
+                    string_types += (bytes,)
+                nan = np.nan
+                if all(isinstance(x, string_types) or x is nan for x in data):
+                    # Force any missing (null) values to b''.  Numpy will
+                    # upcast to str/unicode as needed.
+                    data[mask] = b''
+
+                    # When the numpy object array is represented as a list then
+                    # numpy initializes to the correct string or unicode type.
+                    data = np.array([x for x in data])
+
+            if np.any(mask):
+                out[name] = MaskedColumn(data=data, name=name, mask=mask)
+            else:
+                out[name] = Column(data=data, name=name)
+
+        return cls(out)
+
 
 class QTable(Table):
     """A class to represent tables of heterogeneous data.